Move LanguageRegistry into its own module
